+++ bootique-jdbc-junit5/src/main/java/io/bootique/jdbc/junit5/matcher/CsvMatcherRunner.java
@@ -79,7 +79,8 @@ class CsvMatcherRunner {
     }
 
     private List<Object[]> readTableData() {
-        List<Object[]> data = table.selectColumns(referenceData.getHeader());
+        DbColumnMetadata[] header = referenceData.getHeader();
+        List<Object[]> data = table.selectColumns(header).select();
         assertSizeMatches(referenceData, data);
         return data;
     }