+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #105 Upgrade Testcontainers to 1.15.0-rc2
 * #106 Test API: Table.delete().where(..)
+* #107 Test API - better SELECT builder
 
 ## 2.0.M1